+++ examples/next/pages/ui/components/authenticator/custom-slots/index.page.tsx
@@ -120,11 +120,142 @@ const components = {
       return <Text>Footer Information</Text>;
     },
   },
+  SetupTOTP: {
+    Header() {
+      const { tokens } = useTheme();
+      return (
+        <Heading
+          padding={`${tokens.space.xl} 0 0 ${tokens.space.xl}`}
+          level={3}
+        >
+          Enter Information:
+        </Heading>
+      );
+    },
+    Footer() {
+      return <Text>Footer Information</Text>;
+    },
+  },
+  ConfirmSignIn: {
+    Header() {
+      const { tokens } = useTheme();
+      return (
+        <Heading
+          padding={`${tokens.space.xl} 0 0 ${tokens.space.xl}`}
+          level={3}
+        >
+          Enter Information:
+        </Heading>
+      );
+    },
+    Footer() {
+      return <Text>Footer Information</Text>;
+    },
+  },
+  ResetPassword: {
+    Header() {
+      const { tokens } = useTheme();
+      return (
+        <Heading
+          padding={`${tokens.space.xl} 0 0 ${tokens.space.xl}`}
+          level={3}
+        >
+          Enter Information:
+        </Heading>
+      );
+    },
+    Footer() {
+      return <Text>Footer Information</Text>;
+    },
+  },
+  ConfirmResetPassword: {
+    Header() {
+      const { tokens } = useTheme();
+      return (
+        <Heading
+          padding={`${tokens.space.xl} 0 0 ${tokens.space.xl}`}
+          level={3}
+        >
+          Enter Information:
+        </Heading>
+      );
+    },
+    Footer() {
+      return <Text>Footer Information</Text>;
+    },
+  },
+};
+
+const formFields = {
+  signIn: {
+    username: {
+      labelHidden: false,
+      placeholder: 'Enter your email',
+    },
+  },
+  signUp: {
+    password: {
+      labelHidden: false,
+      label: 'Password:',
+      placeholder: 'Enter your Password:',
+      isRequired: false,
+      order: 2,
+    },
+    confirm_password: {
+      labelHidden: false,
+      label: 'Confirm Password:',
+      order: 1,
+    },
+  },
+  forceNewPassword: {
+    password: {
+      labelHidden: false,
+      placeholder: 'Enter your Password:',
+    },
+  },
+  resetPassword: {
+    username: {
+      labelHidden: false,
+      placeholder: 'Enter your email:',
+    },
+  },
+  confirmResetPassword: {
+    confirmation_code: {
+      labelHidden: false,
+      placeholder: 'Enter your Confirmation Code:',
+      label: 'New Label',
+      isRequired: false,
+    },
+    confirm_password: {
+      labelHidden: false,
+      placeholder: 'Enter your Password Please:',
+    },
+  },
+  setupTOTP: {
+    QR: {
+      totpIssuer: 'test issuer',
+      totpUsername: 'amlify_qr_test_user',
+    },
+    confirmation_code: {
+      labelHidden: false,
+      label: 'New Label',
+      placeholder: 'Enter your Confirmation Code:',
+      isRequired: false,
+    },
+  },
+  confirmSignIn: {
+    confirmation_code: {
+      labelHidden: false,
+      label: 'New Label',
+      placeholder: 'Enter your Confirmation Code:',
+      isRequired: false,
+    },
+  },
 };
 
 export default function App() {
   return (
-    <Authenticator components={components}>
+    <Authenticator formFields={formFields} components={components}>
       {({ signOut }) => <button onClick={signOut}>Sign out</button>}
     </Authenticator>
   );