+++ examples/vue/src/pages/ui/components/authenticator/custom-slots/index.vue
@@ -12,10 +12,77 @@ import awsExports from './aws-exports';
 Amplify.configure(awsExports);
 
 const { toResetPassword, toSignIn } = toRefs(useAuthenticator());
+
+const formFields = {
+  signIn: {
+    username: {
+      labelHidden: false,
+      placeholder: 'Enter your email',
+    },
+  },
+  signUp: {
+    confirm_password: {
+      labelHidden: false,
+      label: 'Confirm Password:',
+      order: 1,
+    },
+    password: {
+      labelHidden: false,
+      label: 'Password:',
+      placeholder: 'Enter your Password:',
+      isRequired: false,
+      order: 2,
+    },
+  },
+  forceNewPassword: {
+    password: {
+      labelHidden: false,
+      placeholder: 'Enter your Password:',
+    },
+  },
+  resetPassword: {
+    username: {
+      labelHidden: false,
+      placeholder: 'Enter your email:',
+    },
+  },
+  confirmResetPassword: {
+    confirmation_code: {
+      labelHidden: false,
+      placeholder: 'Enter your Confirmation Code:',
+      label: 'New Label',
+      isRequired: false,
+    },
+    confirm_password: {
+      labelHidden: false,
+      placeholder: 'Enter your Password Please:',
+    },
+  },
+  setupTOTP: {
+    QR: {
+      totpIssuer: 'test issuer',
+      totpUsername: 'amlify_qr_test_user',
+    },
+    confirmation_code: {
+      labelHidden: false,
+      label: 'New Label',
+      placeholder: 'Enter your Confirmation Code:',
+      isRequired: false,
+    },
+  },
+  confirmSignIn: {
+    confirmation_code: {
+      labelHidden: false,
+      label: 'New Label',
+      placeholder: 'Enter your Confirmation Code:',
+      isRequired: false,
+    },
+  },
+};
 </script>
 
 <template>
-  <authenticator>
+  <authenticator :form-fields="formFields">
     <template v-slot:header>
       <div style="padding: var(--amplify-space-large); text-align: center">
         <img
@@ -94,6 +161,58 @@ const { toResetPassword, toSignIn } = toRefs(useAuthenticator());
     </template>
 
     <template v-slot:confirm-sign-up-footer>
+      <div>Footer Information</div>
+    </template>
+
+    <template v-slot:setup-totp-header>
+      <h3
+        class="amplify-heading"
+        style="padding: var(--amplify-space-xl) 0 0 var(--amplify-space-xl)"
+      >
+        Enter Information:
+      </h3>
+    </template>
+
+    <template v-slot:setup-totp-footer>
+      <div>Footer Information</div>
+    </template>
+
+    <template v-slot:confirm-sign-in-header>
+      <h3
+        class="amplify-heading"
+        style="padding: var(--amplify-space-xl) 0 0 var(--amplify-space-xl)"
+      >
+        Enter Information:
+      </h3>
+    </template>
+
+    <template v-slot:confirm-sign-in-footer>
+      <div>Footer Information</div>
+    </template>
+
+    <template v-slot:reset-password-header>
+      <h3
+        class="amplify-heading"
+        style="padding: var(--amplify-space-xl) 0 0 var(--amplify-space-xl)"
+      >
+        Enter Information:
+      </h3>
+    </template>
+
+    <template v-slot:reset-password-footer>
+      <div>Footer Information</div>
+    </template>
+
+    <template v-slot:confirm-reset-password-header>
+      <h3
+        class="amplify-heading"
+        style="padding: var(--amplify-space-xl) 0 0 var(--amplify-space-xl)"
+      >
+        Enter Information:
+      </h3>
+    </template>
+
+    <template v-slot:confirm-reset-password-footer>
       <div>Footer Information</div>
     </template>