+++ packages/react/src/components/Authenticator/Authenticator.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 import { AuthenticatorMachineOptions } from '@aws-amplify/ui';
 
 import { Provider, useAuthenticator } from './hooks/useAuthenticator';
-import { ResetPassword } from './ResetPassword';
 import { Router, RouterProps } from './Router';
 import { SetupTOTP } from './SetupTOTP';
 import { SignIn } from './SignIn';
 import { SignUp } from './SignUp';
 import { ForceNewPassword } from './ForceNewPassword';
+import { ResetPassword } from './ResetPassword';
 import {
   CustomComponentsContext,
   ComponentsProviderProps,
@@ -31,6 +31,7 @@ export function Authenticator({
   socialProviders,
   variation,
   hideSignUp,
+  formFields,
 }: AuthenticatorProps) {
   const components = { ...defaultComponents, ...customComponents };
   const machineProps = {
@@ -39,6 +40,7 @@ export function Authenticator({
     services,
     signUpAttributes,
     socialProviders,
+    formFields,
   };
 
   // Helper component that sends init event to the parent provider