+++ packages/react/src/components/Authenticator/VerifyUser/VerifyUser.tsx
@@ -11,6 +11,7 @@ import {
 import { useAuthenticator } from '..';
 import { Heading, Radio, RadioGroupField } from '../../..';
 import { RemoteErrorMessage, TwoButtonSubmitFooter } from '../shared';
+import { useCustomComponents } from '../hooks/useCustomComponents';
 import {
   isInputOrSelectElement,
   isInputElement,
@@ -55,6 +56,12 @@ const generateRadioGroup = (
 };
 
 export const VerifyUser = (): JSX.Element => {
+  const {
+    components: {
+      VerifyUser: { Header = VerifyUser.Header, Footer = VerifyUser.Footer },
+    },
+  } = useCustomComponents();
+
   const { _state, isPending, submitForm, updateForm } = useAuthenticator();
   const context = getActorContext(_state) as SignInContext;
 
@@ -108,9 +115,7 @@ export const VerifyUser = (): JSX.Element => {
         className="amplify-flex"
         disabled={isPending}
       >
-        <Heading level={3}>
-          {translate('Account recovery requires verified contact information')}
-        </Heading>
+        <Header />
 
         {verificationRadioGroup}
 
@@ -121,7 +126,18 @@ export const VerifyUser = (): JSX.Element => {
           cancelButtonSendType="SKIP"
           submitButtonText={footerSubmitText}
         />
+        <Footer />
       </fieldset>
     </form>
   );
 };
+
+VerifyUser.Header = () => {
+  return (
+    <Heading level={3}>
+      {translate('Account recovery requires verified contact information')}
+    </Heading>
+  );
+};
+
+VerifyUser.Footer = (): JSX.Element => null;